Move redoMainChainImport to Cli and blockchain
